Time: 41 ms (97.09%), Space: 210.5 MB (31.72%) - LeetHub

+++ 0403-frog-jump/0403-frog-jump.cpp
@@ -7,9 +7,9 @@ public:
         if(dp[ind][prev] != -1)
             return dp[ind][prev];
         
-        long long prev_jump = stones[ind] - stones[prev];
+        int prev_jump = stones[ind] - stones[prev];
         
-        long long max_reach = stones[ind] + prev_jump + 1; 
+        int max_reach = stones[ind] + prev_jump + 1; 
         
         for(int j = ind + 1; j < n and stones[j] <= max_reach; ++j) {
